requirements.txt
Flask==3.0.3
reportlab==4.2.2
Pillow==10.4.0
PyJWT==2.9.0
Flask-Limiter==3.7.0
python-dotenv==1.0.1
gunicorn==22.0.0
